+++ codec/src/lib.rs
@@ -441,7 +441,7 @@ macro_rules! pdu {
 /// The overall version of the codec.
 /// This must be bumped when backwards incompatible changes
 /// are made to the types and protocol.
-pub const CODEC_VERSION: usize = 43;
+pub const CODEC_VERSION: usize = 44;
 
 // Defines the Pdu enum.
 // Each struct has an explicit identifying number.