+++ docs/changelog.md
@@ -58,6 +58,9 @@ As features stabilize some brief notes about them will accumulate here.
   that WezTerm.app be code-signed, which is the case for official binaries.
 * [ShowLauncherArgs](config/lua/keyassignment/ShowLauncherArgs.md) now allows
   customizing the help text. Thanks to @mgpinf! #6606
+* Preliminary support for ConEmu style progress escape sequences. See
+  [pane:get_progress()](config/lua/pane/get_progress.md) for more information.
+  #6581
 
 #### New
 * [wezterm.serde](config/lua/wezterm.serde/index.md) module for serialization